Alle permissies toegevoegd en getest

+++ api/views/score.py
@@ -3,14 +3,22 @@ from rest_framework.response import Response
 from rest_framework import status
 
 from api.models.score import Score
+from api.models.groep import Groep
+from api.models.indiening import Indiening
 from api.serializers.score import ScoreSerializer
+from api.utils import is_lesgever, contains
 
 
 @api_view(['GET', 'POST'])
 def score_list(request, format=None):
 
     if request.method == 'GET':
-        scores = Score.objects.all()
+        if is_lesgever(request.user):
+            scores = Score.objects.all()
+        else:
+            groepen = Groep.objects.filter(studenten=request.user.id)
+            indieningen = Indiening.objects.filter(groep__in=groepen)
+            scores = Score.objects.filter(indiening__in=indieningen)
 
         if "indiening" in request.GET:
             try:
@@ -23,11 +31,13 @@ def score_list(request, format=None):
         return Response(serializer.data)
     
     elif request.method == 'POST':
-        serializer = ScoreSerializer(data=request.data)
-        if serializer.is_valid():
-            serializer.save()
-            return Response(serializer.data, status=status.HTTP_201_CREATED)
-        return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+        if is_lesgever(request.user):
+            serializer = ScoreSerializer(data=request.data)
+            if serializer.is_valid():
+                serializer.save()
+                return Response(serializer.data, status=status.HTTP_201_CREATED)
+            return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+        return Response(status=status.HTTP_403_FORBIDDEN)
         
 @api_view(['GET', 'PUT', 'DELETE'])
 def score_detail(request, id, format=None): 
@@ -35,18 +45,22 @@ def score_detail(request, id, format=None):
         score = Score.objects.get(pk=id)
     except Score.DoesNotExist:
         return Response(status=status.HTTP_404_NOT_FOUND)
-
+    
     if request.method == 'GET':
-        serializer = ScoreSerializer(score)
-        return Response(serializer.data)
-    
-    elif request.method == 'PUT':
-        serializer = ScoreSerializer(score, data=request.data)
-        if serializer.is_valid():
-            serializer.save()
+        if is_lesgever(request.user) or contains(score.indiening.groep.studenten, request.user):
+            serializer = ScoreSerializer(score)
             return Response(serializer.data)
-        return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
-    
-    elif request.method == 'DELETE':
-        score.delete()
-        return Response(status=status.HTTP_204_NO_CONTENT)
+        return Response(status=status.HTTP_403_FORBIDDEN)
+
+    if is_lesgever(request.user): 
+        if request.method == 'PUT':
+            serializer = ScoreSerializer(score, data=request.data)
+            if serializer.is_valid():
+                serializer.save()
+                return Response(serializer.data)
+            return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+        
+        elif request.method == 'DELETE':
+            score.delete()
+            return Response(status=status.HTTP_204_NO_CONTENT)
+    return Response(status=status.HTTP_403_FORBIDDEN)
